+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Pomniejsze usprawnienia do TimerHeap
 * Dodano `self` do @typed
+* Dodano typ `Number` do @typed
 
 ## v2.0.11